Fix: stored procedure not working correctly

+++ src/API/API.Infrastructure/Repositories/UserRepository.cs
@@ -41,10 +41,10 @@ namespace API.Infrastructure.Repositories
             {
                 var parameters = new DynamicParameters();
                 parameters.Add("@userId", userId);
-                parameters.Add("@sex", userParams.Sex);
+                parameters.Add("@sex", (int)userParams.Sex);
                 parameters.Add("@minAge", userParams.MinAge);
                 parameters.Add("@maxAge", userParams.MaxAge);
-                parameters.Add("@orderBy", userParams.OrderBy);
+                parameters.Add("@orderBy", (int)userParams.OrderBy);
                 parameters.Add("@pageNumber", userParams.PageNumber);
                 parameters.Add("@pageSize", userParams.ItemsPerPage);
